# apps/backend/src/modules/knowledge/chat/message/message.controller.ts
import { MessageEntity } from '@/entities/message.entity';
import { BaseController } from '@/shared/controller';
import { Controller, HttpPost } from '@/shared/controller/decorators';
import { Body } from '@nestjs/common';
import { ApiResponse } from '@nestjs/swagger';

import { CreateMessageDto, CreateMessageResponseDto } from './dto/create-message.dto';
import { MessageService } from './message.service';

@Controller('tenant/:tenantId/knowledge/:knowledgeSlug/chat/:chatId/message')
export class MessageController extends BaseController({ entity: MessageEntity }) {
  constructor(public service: MessageService) {
    super(service);
  }

  @HttpPost("new")
  @ApiResponse({ type: CreateMessageResponseDto, status: 200 })
  async newMessage(@Body() body: CreateMessageDto) {
    return this.service.sendMessage(body.content);
  }
}