# apps/backend/swagger.yaml
openapi: 3.0.0
paths:
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/source/{id}/view":
    get:
      operationId: SourceController_view
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileURLResponseDto"
      security:
        - access-token: []
      tags:
        &a1
        - Source
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/source/{id}/download-url":
    get:
      operationId: SourceController_download
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileURLResponseDto"
      security:
        - access-token: []
      tags: *a1
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/source/upload-url":
    post:
      operationId: SourceController_upload
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetUploadUrlDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileURLResponseDto"
      security:
        - access-token: []
      tags: *a1
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/source/{id}/retry":
    post:
      operationId: SourceController_retry
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceEntity"
      security:
        - access-token: []
      tags: *a1
...
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/source":
    get:
      operationId: SourceController_findMany
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Records found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceEntity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - access-token: []
      tags: *a1
...
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: ""
      tags:
        &a4
        - Auth
...
  /api/auth/create-account:
    post:
      operationId: AuthController_createAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountDto"
      responses:
        "201":
          description: ""
      tags: *a4
  /api/auth/active-account:
    post:
      operationId: AuthController_activeAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActiveAccountDto"
      responses:
        "201":
          description: ""
      tags: *a4
  /api/auth/forget-password:
    post:
      operationId: AuthController_forgetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgetPasswordDto"
      responses:
        "201":
          description: ""
      tags: *a4
...
  "/api/tenant/{tenantId}/invite/send":
    post:
      operationId: InviteController_send
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInviteDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SendInviteResponseDto"
      security:
        - access-token: []
      tags:
        &a11
        - Invite
...
  "/api/tenant/{tenantId}/invite":
    get:
      operationId: InviteController_findMany
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Records found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InviteEntity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - access-token: []
      tags: *a11
    post:
      operationId: InviteController_create
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: Record created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteEntity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - duplicate record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - access-token: []
      tags: *a11
...
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/chat/{chatId}/message/new":
    post:
      operationId: MessageController_newMessage
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMessageResponseDto"
      security:
        - access-token: []
      tags:
        &a13
        - Message
...
  "/api/tenant/{tenantId}/knowledge/{knowledgeSlug}/chat/{chatId}/message":
    get:
      operationId: MessageController_findMany
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Records found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageEntity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - access-token: []
      tags: *a13
    post:
      operationId: MessageController_create
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: knowledgeSlug
          in: path
          required: true
          schema:
            type: string
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: Record created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEntity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - duplicate record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - access-token: []
      tags: *a13
...
info:
  title: Snipet
  description: The Snipet API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Enter JWT token
  schemas:
    FileURLResponseDto:
      type: object
      properties:
        url:
          type: string
          example: https://example.com/file.pdf
        key:
          type: string
          example: path/to/my-file.pdf
      required:
        - url
        - key
    GetUploadUrlDto:
      type: object
      properties:
        fileName:
          type: string
          example: my-file.pdf
        contentType:
          type: string
          example: application/pdf
        fileSize:
          type: number
          example: 104857600
          description: "File size in bytes (max: 100.00 MB"
      required:
        - fileName
        - contentType
        - fileSize
    InviteEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the invite
          format: uuid
        tenantId:
          type: string
          description: The ID of the tenant this invite belongs to
          format: uuid
        email:
          type: string
          description: The email address of the person being invited
          format: email
        roleId:
          type: string
          description: The ID of the role assigned to the invited person
          format: uuid
        userId:
          type: string
          description: The ID of the user being invited, if they already exist
          format: uuid
        creatorId:
          type: string
          description: The ID of the user who created the invite
          format: uuid
        expiresAt:
          format: date-time
          type: string
          description: The timestamp when the invite expires
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the invite was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the invite was last updated
        tenant:
          $ref: "#/components/schemas/TenantEntity"
        role:
          $ref: "#/components/schemas/RoleEntity"
        user:
          $ref: "#/components/schemas/UserEntity"
        creator:
          $ref: "#/components/schemas/MemberEntity"
      required:
        - id
        - tenantId
        - email
        - roleId
        - creatorId
        - expiresAt
        - createdAt
        - updatedAt
        - tenant
        - role
        - creator
    RoleEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the role
          format: uuid
        key:
          type: string
          description: The key of the role
        name:
          type: string
          description: The name of the role
        permissions:
          type: number
          description: The permissions of the role as a bitmask
        scope:
          type: string
          description: The scope of the role
          enum:
            - TENANT
            - GLOBAL
        tenantId:
          type: string
          description: The ID of the tenant this role belongs to
          format: uuid
        createdById:
          type: string
          description: The ID of the user who created the role
          format: uuid
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the role was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the role was last updated
        tenant:
          $ref: "#/components/schemas/TenantEntity"
        createdBy:
          $ref: "#/components/schemas/MemberEntity"
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserEntity"
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberEntity"
        invites:
          type: array
          items:
            $ref: "#/components/schemas/InviteEntity"
      required:
        - id
        - key
        - name
        - permissions
        - scope
        - createdAt
        - updatedAt
...
    MemberEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the member
          format: uuid
        owner:
          type: boolean
          description: Indicates if the member is the owner of the tenant
          default: false
        userId:
          type: string
          description: The ID of the user
          format: uuid
        tenantId:
          type: string
          description: The ID of the tenant
          format: uuid
        roleId:
          type: string
          description: The ID of the role assigned to the member
          format: uuid
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the member was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the member was last updated
        user:
          $ref: "#/components/schemas/UserEntity"
        tenant:
          $ref: "#/components/schemas/TenantEntity"
        role:
          $ref: "#/components/schemas/RoleEntity"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationEntity"
        invites:
          type: array
          items:
            $ref: "#/components/schemas/InviteEntity"
      required:
        - id
        - owner
        - userId
        - tenantId
        - roleId
        - createdAt
        - updatedAt
    TenantEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the tenant
          format: uuid
        name:
          type: string
          description: The name of the tenant
        description:
          type: string
          description: A brief description of the tenant
        backgroundImage:
          type: string
          description: The URL of the background image for the tenant
          format: uri
        disabledAt:
          format: date-time
          type: string
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the tenant was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the tenant was last updated
        members:
          description: The members associated with this tenant
          type: array
          items:
            $ref: "#/components/schemas/MemberEntity"
        invites:
          description: The invites associated with this tenant
          type: array
          items:
            $ref: "#/components/schemas/InviteEntity"
        notifications:
          description: The notifications associated with this tenant
          type: array
          items:
            $ref: "#/components/schemas/NotificationEntity"
        roles:
          description: The roles defined within this tenant
          type: array
          items:
            $ref: "#/components/schemas/RoleEntity"
        sources:
          description: The sources associated with this tenant
          type: array
          items:
            $ref: "#/components/schemas/SourceEntity"
      required:
        - id
        - name
        - backgroundImage
        - createdAt
        - updatedAt
    SourceEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the source
          format: uuid
        key:
          type: string
          description: The key of the source file in the storage
        path:
          type: string
          description: The path of the source file
        name:
          type: string
          description: The name of the source
        description:
          type: string
          description: A brief description of the source
        originalName:
          type: string
          description: The original name of the file
        metadata:
          type: object
          description: The metadata of the source
        sourceType:
          type: string
          description: The type of the source
          enum:
            - TEXT
            - DOC
            - LINK
            - VIDEO
            - AUDIO
            - IMAGE
        indexStatus:
          type: string
          description: The indexing status of the source
          enum:
            - PENDING
            - INDEXING
            - INDEXED
            - ERROR
        indexError:
          type: string
          description: The error message if indexing fails
        memoryId:
          type: string
          description: The ID of the memory associated with this source
          format: uuid
        knowledgeId:
          type: string
          description: The ID of the knowledge base this source belongs to
          format: uuid
        tenantId:
          type: string
          description: The ID of the tenant this source belongs to
          format: uuid
        folderId:
          type: string
          description: The ID of the folder this source belongs to
          format: uuid
        knowledge:
          $ref: "#/components/schemas/KnowledgeEntity"
        folder:
          $ref: "#/components/schemas/FolderEntity"
        tenant:
          $ref: "#/components/schemas/TenantEntity"
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the source was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the source was last updated
      required:
        - id
        - key
        - path
        - name
        - metadata
        - sourceType
        - indexStatus
        - knowledgeId
        - tenantId
        - createdAt
        - updatedAt
    FolderEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the folder
          format: uuid
        knowledgeId:
          type: string
          description: The ID of the associated knowledge base
          format: uuid
        knowledge:
          $ref: "#/components/schemas/KnowledgeEntity"
        name:
          type: string
          description: The name of the folder
        root:
          type: boolean
          description: Indicates if the folder is a root folder
        parentId:
          type: string
          description: The ID of the parent folder
          format: uuid
        parent:
          $ref: "#/components/schemas/FolderEntity"
        children:
          type: array
          items:
            $ref: "#/components/schemas/FolderEntity"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceEntity"
        tenantId:
          type: string
          description: The ID of the tenant this folder belongs to
          format: uuid
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the folder was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the folder was last updated
      required:
        - id
        - knowledgeId
        - knowledge
        - name
        - children
        - sources
        - tenantId
        - createdAt
        - updatedAt
    MessageEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the message
          format: uuid
        messageRole:
          type: string
          description: The role of the message sender
          enum:
            - user
            - assistant
            - system
        content:
          type: string
          description: The content of the message
        chatId:
          type: string
          description: The ID of the chat this message belongs to
          format: uuid
        chat:
          $ref: "#/components/schemas/ChatEntity"
        knowledgeId:
          type: string
          description: The ID of the knowledge base associated with this message
          format: uuid
        knowledge:
          $ref: "#/components/schemas/KnowledgeEntity"
        tenantId:
          type: string
          description: The ID of the tenant this message belongs to
          format: uuid
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the message was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the message was last updated
      required:
        - id
        - messageRole
        - content
        - chatId
        - knowledgeId
        - tenantId
        - createdAt
        - updatedAt
    ChatEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the chat
          format: uuid
        name:
          type: string
          description: The name of the chat
        knowledgeId:
          type: string
          description: The ID of the associated knowledge base
          format: uuid
        knowledge:
          $ref: "#/components/schemas/KnowledgeEntity"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageEntity"
        tenantId:
          type: string
          description: The ID of the tenant this chat belongs to
          format: uuid
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the chat was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the chat was last updated
      required:
        - id
        - name
        - knowledgeId
        - knowledge
        - messages
        - tenantId
        - createdAt
        - updatedAt
...
    KnowledgeEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the knowledge base
          format: uuid
        slug:
          type: string
          description: The URL-friendly slug for the knowledge base
        title:
          type: string
          description: The title of the knowledge base
        description:
          type: string
          description: A brief description of the knowledge base
        status:
          type: string
          description: The status of the knowledge base
          enum:
            - DELETING
            - DELETE_ERROR
            - OK
        deleteError:
          type: string
          description: The error message if deletion fails
        files:
          type: number
          description: The number of files in the knowledge base
        storage:
          type: number
          description: The total storage used by the knowledge base in bytes
        tenantId:
          type: string
          description: The ID of the tenant this knowledge base belongs to
          format: uuid
        createdAt:
          format: date-time
          type: string
          description: The timestamp when the knowledge base was created
        updatedAt:
          format: date-time
          type: string
          description: The timestamp when the knowledge base was last updated
        folders:
          type: array
          items:
            $ref: "#/components/schemas/FolderEntity"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceEntity"
        chats:
          type: array
          items:
            $ref: "#/components/schemas/ChatEntity"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageEntity"
        knowledgeLLMs:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeLLMEntity"
      required:
        - id
        - slug
        - title
        - status
        - files
        - storage
        - tenantId
        - createdAt
        - updatedAt
        - folders
        - sources
        - chats
        - messages
        - knowledgeLLMs
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        timestamp:
          format: date-time
          type: string
      required:
        - error
        - code
        - timestamp
...
    LoginDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        redirect:
          type: string
      required:
        - email
        - password
    CreateAccountDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
          maxLength: 100
      required:
        - name
        - email
        - password
    ActiveAccountDto:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    ForgetPasswordDto:
      type: object
      properties:
        email:
          type: string
      required:
        - email
...
    SendInviteItemDto:
      type: object
      properties:
        email:
          type: string
        roleId:
          type: string
      required:
        - email
        - roleId
    SendInviteDto:
      type: object
      properties:
        emails:
          type: array
          items:
            $ref: "#/components/schemas/SendInviteItemDto"
      required:
        - emails
    SendInviteResponseDto:
      type: object
      properties:
        sendedInvites:
          type: array
          items:
            $ref: "#/components/schemas/InviteEntity"
        reSendedInvites:
          type: array
          items:
            $ref: "#/components/schemas/InviteEntity"
        alreadyInTenant:
          type: array
          items:
            $ref: "#/components/schemas/MemberEntity"
      required:
        - sendedInvites
        - reSendedInvites
        - alreadyInTenant
...
    CreateMessageDto:
      type: object
      properties:
        content:
          type: string
          description: The content of the message
      required:
        - content
    CreateMessageResponseDto:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/MessageEntity"
        aiMessage:
          $ref: "#/components/schemas/MessageEntity"
      required:
        - userMessage
        - aiMessage
...